Move AdminController to the integration as ScheduledActionManager
